Feature/#103 detail plan register

+++ src/main/java/com/ssafy/tourmates/admin/attraction/AttractionInfo.java
@@ -16,7 +16,7 @@ public class AttractionInfo {
 
     @Id
     @Column(name = "content_id")
-    private Integer id;
+    private Long id;
     @Column(length = 100)
     private String title;
     @Column(length = 100)
@@ -47,7 +47,7 @@ public class AttractionInfo {
     private Gugun gugun;
 
     @Builder
-    public AttractionInfo(Integer id, String title, String addr1, String addr2, String zipcode, String tel, String firstImage, String firstImage2, int readcount, double latitude, double longitude, String mlevel, Sido sido, Gugun gugun) {
+    public AttractionInfo(Long id, String title, String addr1, String addr2, String zipcode, String tel, String firstImage, String firstImage2, int readcount, double latitude, double longitude, String mlevel, Sido sido, Gugun gugun) {
         this.id = id;
         this.title = title;
         this.addr1 = addr1;
